delete service abstraction
